requirements.txt
streamlit
snowflake-connector-python
pandas
plotly
sqlalchemy
python-dotenv 